+++ operator/cloudscale/client.go
@@ -6,15 +6,10 @@ import (
 
 	pipeline "github.com/ccremer/go-command-pipeline"
 	cloudscalesdk "github.com/cloudscale-ch/cloudscale-go-sdk/v2"
-	bucketv1 "github.com/vshn/appcat-service-s3/apis/bucket/v1"
 	cloudscalev1 "github.com/vshn/appcat-service-s3/apis/cloudscale/v1"
 	"github.com/vshn/appcat-service-s3/operator/steps"
 	"golang.org/x/oauth2"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	controllerruntime "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 // APIToken is the authentication token to use against cloudscale.ch API
@@ -50,7 +45,7 @@ func CreateObjectsUser(ctx context.Context) error {
 	user.Status.UserID = csUser.ID
 
 	pipeline.StoreInContext(ctx, CloudscaleUserKey{}, csUser)
-	return logIfNotError(err, log, 1, "Created objects user in cloudscale")
+	return logIfNotError(err, log, 1, "Created objects user in cloudscale", "userID", csUser.ID)
 }
 
 // GetObjectsUser fetches an existing objects user from the project associated with the API token.
@@ -63,39 +58,6 @@ func GetObjectsUser(ctx context.Context) error {
 
 	pipeline.StoreInContext(ctx, CloudscaleUserKey{}, csUser)
 	return logIfNotError(err, log, 1, "Fetched objects user in cloudscale")
-}
-
-// UserCredentialSecretKey identifies the credential Secret in the context.
-type UserCredentialSecretKey struct{}
-
-// EnsureCredentialSecret creates the credential secret.
-func EnsureCredentialSecret(ctx context.Context) error {
-	kube := steps.GetClientFromContext(ctx)
-	user := steps.GetFromContextOrPanic(ctx, ObjectsUserKey{}).(*cloudscalev1.ObjectsUser)
-	csUser := steps.GetFromContextOrPanic(ctx, CloudscaleUserKey{}).(*cloudscalesdk.ObjectsUser)
-	log := controllerruntime.LoggerFrom(ctx)
-
-	secret := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: user.Spec.SecretRef, Namespace: user.Namespace}}
-
-	if keyErr := checkUserForKeys(csUser); keyErr != nil {
-		return keyErr
-	}
-
-	// See https://www.cloudscale.ch/en/api/v1#objects-users
-
-	_, err := controllerruntime.CreateOrUpdate(ctx, kube, secret, func() error {
-		secret.Labels = labels.Merge(secret.Labels, getCommonLabels(user.Name))
-		if secret.StringData == nil {
-			secret.StringData = map[string]string{}
-		}
-		secret.StringData[bucketv1.AccessKeyIDName] = csUser.Keys[0]["access_key"]
-		secret.StringData[bucketv1.SecretAccessKeyName] = csUser.Keys[0]["secret_key"]
-		controllerutil.AddFinalizer(secret, userFinalizer)
-		return controllerutil.SetOwnerReference(user, secret, kube.Scheme())
-	})
-
-	pipeline.StoreInContext(ctx, UserCredentialSecretKey{}, secret)
-	return logIfNotError(err, log, 1, "Ensured credential secret", "secretName", user.Spec.SecretRef)
 }
 
 func checkUserForKeys(user *cloudscalesdk.ObjectsUser) error {
